// src/modules/land-works/components/scene-builder-card/types.ts
export interface NotionResultForCard {
  coverPhotoLink: string;
  avatarPhotoLink: string;
  builderName: string;
  definition: string;
  builderType: string;
  shortDescription: string;
  location: string;
}